requirements.txt
Flask
mysql-connector-python
googletrans==4.0.0-rc1
Flask-Cors
